database: Rename DB module to Database
